Created Player and Enemy subclasses
