+++ src/asset/sprite.py
@@ -1,7 +1,5 @@
 import pygame
 import time
-import random
-from collections import deque
 from rect.utils import define_rect
 
 
@@ -130,60 +128,6 @@ class Sprite:
                 if maze_grid[row][col] == 1:
                     return False
         return True
-
-    # Method to check if the sprite can move within maze
-    # in the direction of a second sprite, and that a clear
-    # path exists between the two. If true, returns the direction
-    # to be traversed by first sprite.
-    def can_move_towards(self, second_sprite: "Sprite", maze_grid):
-        aligned_horz = False
-        aligned_vert = False
-        # Check if aligned horizontally with second sprite (same y)
-        if self.center_position[1] == second_sprite.center_position[1]:
-            # Move towards second sprite either right or left
-            aligned_horz = True
-            if self.center_position[0] < second_sprite.center_position[0]:
-                move_direction = (1, 0)
-            elif self.center_position[0] > second_sprite.center_position[0]:
-                move_direction = (-1, 0)
-        # Otherwise check if aligned vertically with second sprite (same x)
-        elif self.center_position[0] == second_sprite.center_position[0]:
-            # Move towards second sprite either up or down
-            aligned_vert = True
-            if self.center_position[1] > second_sprite.center_position[1]:
-                move_direction = (0, -1)
-            elif self.center_position[1] < second_sprite.center_position[1]:
-                move_direction = (0, 1)
-
-        # Form a horizontal rect between the two for checking
-        if aligned_horz:
-            temp_topleft_x = min(
-                self.path_rect.topleft[0], second_sprite.path_rect.topleft[0]
-            ) + int(self.path_width / 2)
-            temp_topleft_y = self.path_rect.topleft[1]
-            temp_width = abs(
-                self.path_rect.center[0] - second_sprite.path_rect.center[0]
-            )
-            temp_height = self.path_rect.height
-        # Form a vertical rect between the two for checking
-        elif aligned_vert:
-            temp_topleft_x = self.path_rect.topleft[0]
-            temp_topleft_y = min(
-                self.path_rect.topleft[1], second_sprite.path_rect.topleft[1]
-            ) + int(self.path_width / 2)
-            temp_width = self.path_rect.width
-            temp_height = abs(
-                self.path_rect.center[1] - second_sprite.path_rect.center[1]
-            )
-
-        if not (aligned_horz or aligned_vert):
-            return ()
-
-        # Check if temp rect resides within a clear space within maze path
-        temp_rect = pygame.Rect(temp_topleft_x, temp_topleft_y, temp_width, temp_height)
-        if Sprite.is_path_clear(temp_rect, maze_grid):
-            return move_direction
-        return ()
 
     # Determine and perform move based on direction, move check, time, and speed
     def perform_move(self, maze_grid, game_tick):
@@ -337,76 +281,9 @@ class Sprite:
     def set_direction(self, x_direction, y_direction):
         self.direction = (x_direction, y_direction)
 
-    # Move towards player if direction possible,
-    # otherwise random navigation, only backtracking if needed
-    def set_navigate_direction(self, second_sprite: "Sprite", maze_grid, game_tick):
-        # Determine how far item has traveled in the game tick
-        if self.move_dist + self.speed * game_tick >= 1:
-            # Possible directions: Right, Left, Up, Down
-            dirs = [(1, 0), (-1, 0), (0, -1), (0, 1)]
-
-            # Check if possible to move towards player with clear path
-            move_direction = ()
-            aligned_second_sprite = False
-            if second_sprite:
-                move_direction = self.can_move_towards(second_sprite, maze_grid)
-            if move_direction:
-                aligned_second_sprite = True
-                self.set_desired_direction(move_direction[0], move_direction[1])
-
-            # If current direction is 0, 0, choose a random direction
-            if self.desired_direction == (0, 0) or not self.desired_direction:
-                rand_direction = dirs[random.randint(0, 3)]
-                self.set_desired_direction(rand_direction[0], rand_direction[1])
-
-            # Check current direction, then orthogonal directions, then reverse
-            reverse_direction = (
-                -1 * self.desired_direction[0],
-                -1 * self.desired_direction[1],
-            )
-            orthog_direction_1 = (self.desired_direction[1], self.desired_direction[0])
-            orthog_direction_2 = (
-                -1 * self.desired_direction[1],
-                -1 * self.desired_direction[0],
-            )
-
-            directions_to_check = [
-                self.desired_direction,
-                orthog_direction_1,
-                orthog_direction_2,
-                reverse_direction,
-            ]
-            valid_indices = []
-            for index, direction in enumerate(directions_to_check):
-                if self.can_move(direction[0], direction[1], maze_grid):
-                    valid_indices.append(index)
-
-            # If no directions are possible or only reverse possible, these take precedent.
-            # Otherwise, pick a random direction of the remaining possibilities if not
-            # aligned with the second sprite (go in player direction if it's possible).
-            if not valid_indices:
-                self.set_desired_direction(0, 0)
-            elif valid_indices == [3]:
-                self.set_desired_direction(
-                    directions_to_check[3][0], directions_to_check[3][1]
-                )
-            elif aligned_second_sprite and 0 in valid_indices:
-                pass
-            else:
-                non_reverse_dirs = [i for i in valid_indices if i != 3]
-                chosen_index = random.choice(non_reverse_dirs)
-                self.set_desired_direction(
-                    directions_to_check[chosen_index][0],
-                    directions_to_check[chosen_index][1],
-                )
-
     # Perform collision check with another sprite
     def collide_check(self, second_sprite: "Sprite"):
         return self.hitbox_rect.colliderect(second_sprite.hitbox_rect)
-
-    # Set sprite speed to input value
-    def set_speed(self, new_speed):
-        self.speed = new_speed
 
     # Set sprite image to original, and reset animation
     def reset_image(self):
@@ -430,14 +307,6 @@ class Sprite:
     # Return flag of whether sprite respawn active
     def can_spawn(self):
         return self.spawn
-
-    # Return rect containing path box
-    def get_path_rect(self):
-        return self.path_rect
-
-    # Return rect containing hit box
-    def get_hit_rect(self):
-        return self.hitbox_rect
 
     # Return center position coordinates
     def get_center_position(self):
